backend/requirements.txt
fastapi>=0.109,<1
uvicorn[standard]>=0.27,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.29,<1
psycopg[binary]>=3.1,<4
alembic>=1.13,<2
pydantic-settings>=2.0,<3
email-validator>=2.0,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=4.0,<4.1

pytest>=8.0,<9
aiosqlite>=0.19,<1
pytest-asyncio>=0.23,<1
httpx>=0.27,<1
aiosmtplib>=3.0,<4
itsdangerous>=2.1,<3
